dk64: split rendering and particles from scenes
