+++ src/DonkeyKong64/scenes.ts
@@ -3,100 +3,46 @@ import * as Viewer from '../viewer.js';
 import * as BYML from '../byml.js';
 import * as UI from '../ui.js';
 
-import { GfxDevice, GfxCullMode, GfxProgram, GfxMegaStateDescriptor, makeTextureDescriptor2D, GfxFormat, GfxSampler, GfxTexture, GfxTexFilterMode, GfxMipFilterMode, GfxBindingLayoutDescriptor, GfxBlendMode, GfxBlendFactor, GfxBuffer, GfxInputLayout, GfxBufferUsage, GfxBufferFrequencyHint, GfxVertexAttributeDescriptor, GfxInputLayoutBufferDescriptor, GfxVertexBufferFrequency, GfxVertexBufferDescriptor, GfxIndexBufferDescriptor } from '../gfx/platform/GfxPlatform.js';
+import { GfxDevice } from '../gfx/platform/GfxPlatform.js';
 import { SceneContext } from '../SceneBase.js';
 import { makeBackbufferDescSimple, standardFullClearRenderPassDescriptor } from '../gfx/helpers/RenderGraphHelpers.js';
-import { F3DEX_Program } from '../BanjoKazooie/render.js';
-import { nArray, align, assert } from '../util.js';
-import { hexzero } from '../util.js';
-import { DeviceProgram } from '../Program.js';
+import { assert, hexzero } from '../util.js';
 import { mat4, vec3 } from 'gl-matrix';
 import { GfxRenderCache } from '../gfx/render/GfxRenderCache.js';
-import { TextureMapping, FakeTextureHolder } from '../TextureHolder.js';
-import { DrawCall, RSP_Geometry, RSPState, runDL_F3DEX2, RSPOutput } from './f3dex2.js';
+import { FakeTextureHolder } from '../TextureHolder.js';
+import { RSPState, runDL_F3DEX2 } from './f3dex2.js';
 import { AnimatedTexture, RSPSharedOutput } from './f3dex2.js';
-import { translateBlendMode, translateCullMode } from '../PokemonSnap/f3dex2.js';
-import { GfxRenderInstList, GfxRenderInstManager } from '../gfx/render/GfxRenderInstManager.js';
-import { GfxRendererLayer, makeSortKey } from '../gfx/render/GfxRenderInstManager.js';
-import { computeViewMatrixSkybox, computeViewMatrix, CameraController } from '../Camera.js';
-import { fillMatrix4x3, fillMatrix4x2, fillVec4, fillMatrix4x4 } from '../gfx/helpers/UniformBufferHelpers.js';
-import { translateCM, Texture, OtherModeH_Layout, OtherModeH_CycleType } from '../Common/N64/RDP.js';
+import { GfxRenderInstList } from '../gfx/render/GfxRenderInstManager.js';
+import { CameraController } from '../Camera.js';
 import { GfxRenderHelper } from '../gfx/render/GfxRenderHelper.js';
-import { TextFilt, ImageFormat, ImageSize } from "../Common/N64/Image.js";
-import { RSPSharedOutput as BanjoRSPSharedOutput, Vertex } from '../BanjoKazooie/f3dex.js';
-import { setAttachmentStateSimple } from '../gfx/helpers/GfxMegaStateDescriptorHelpers.js';
-import { Vec3UnitY, Vec3Zero } from '../MathHelpers.js';
-import { scaleMatrix, setMatrixTranslation } from '../MathHelpers.js';
 
 import ArrayBufferSlice from '../ArrayBufferSlice.js';
 import * as Deflate from '../Common/Compression/Deflate.js';
-import { calcTextureMatrixFromRSPState } from '../Common/N64/RSP.js';
 import { GfxrAttachmentSlot } from '../gfx/render/GfxRenderGraph.js';
-import { createBufferFromData } from '../gfx/helpers/BufferHelpers.js';
-import { ActiveLightCache, buildDynamicLights, buildMapChunkLighting, buildObjectLighting, buildObjectLightingEnvironment, sampleObjectLighting, updateDynamicLighting } from './light.js';
-import type { DynamicLight, DynamicLighting, ObjectLighting, ObjectLightingEnvironment } from './light.js';
+import { ActiveLightCache, buildDynamicLights, buildMapChunkLighting, buildObjectLighting, buildObjectLightingEnvironment } from './light.js';
+import type { DynamicLight, ObjectLightingEnvironment } from './light.js';
 import { ActorAnimationPose, actorModelScale, buildActorMesh, getActorRenderDefinition } from './actor.js';
-import type { ActorAnimationState, ActorMesh, ActorRenderDefinition } from './actor.js';
+import type { ActorRenderDefinition } from './actor.js';
 import { addModel2Props, buildTerrainTriangles } from './prop.js';
-import type { PropAnimationState } from './prop.js';
 import {
     getGeneratedSurfaceAnimatedTextureBindings,
     getSceneNodeAnimatedTextureBindings, initDL, initGeneratedSurfaceMaterial,
-    initSceneNodeMaterial, initSpriteMaterial,
+    initSceneNodeMaterial,
 } from './material.js';
 import type { AnimatedMaterialTextureBinding } from './material.js';
 import { DK64Map, parseInstanceScripts, parseSetup } from './parse.js';
-import type { GeneratedSurface, InstanceScript, ScriptBlock, SetupActor, SetupProp } from './parse.js';
+import type { GeneratedSurface, SetupActor } from './parse.js';
 import { createBackdropRenderer } from './background.js';
 import type { BackdropData, BackdropRenderer } from './background.js';
-import { AABB } from '../Geometry.js';
-import { MeshBounds } from './cull.js';
+import {
+    bindingLayouts, fogPositionToViewDistance, generatedSurfaceHeight, GPUTextureCache,
+    MeshData, RootMeshRenderer, SceneRenderLayer,
+} from './render.js';
+import type { FogParams, Mesh } from './render.js';
+import { addEnvironmentalEffects } from './particles.js';
+import type { EnvironmentParticleData, SpriteData } from './particles.js';
 
 const pathBase = `DonkeyKong64`;
-
-function translateTexture(device: GfxDevice, texture: Texture): GfxTexture {
-    const gfxTexture = device.createTexture(makeTextureDescriptor2D(GfxFormat.U8_RGBA_NORM, texture.width, texture.height, 1));
-    device.setResourceName(gfxTexture, texture.name);
-    device.uploadTextureData(gfxTexture, 0, [texture.pixels]);
-    return gfxTexture;
-}
-
-class GPUTextureCache {
-    private textures = new Map<Texture, GfxTexture>();
-
-    public getTexture(device: GfxDevice, texture: Texture): GfxTexture {
-        let gfxTexture = this.textures.get(texture);
-        if (gfxTexture === undefined) {
-            gfxTexture = translateTexture(device, texture);
-            this.textures.set(texture, gfxTexture);
-        }
-        return gfxTexture;
-    }
-
-    public destroy(device: GfxDevice): void {
-        for (const texture of this.textures.values())
-            device.destroyTexture(texture);
-        this.textures.clear();
-    }
-}
-
-function translateSampler(cache: GfxRenderCache, texture: Texture, linear: boolean): GfxSampler {
-    return cache.createSampler({
-        wrapS: translateCM(texture.tile.cms),
-        wrapT: translateCM(texture.tile.cmt),
-        minFilter: linear ? GfxTexFilterMode.Bilinear : GfxTexFilterMode.Point,
-        magFilter: linear ? GfxTexFilterMode.Bilinear : GfxTexFilterMode.Point,
-        mipFilter: GfxMipFilterMode.Nearest,
-        minLOD: 0, maxLOD: 0,
-    });
-}
-
-function renderModeIsTranslucent(megaStateFlags: Partial<GfxMegaStateDescriptor>): boolean {
-    const blendState = megaStateFlags.attachmentsState?.[0]?.rgbBlendState;
-    return blendState !== undefined
-        && (blendState.blendSrcFactor !== GfxBlendFactor.One
-            || blendState.blendDstFactor !== GfxBlendFactor.Zero);
-}
 
 function resolveAnimatedMaterialTextures(bindings: readonly AnimatedMaterialTextureBinding[], textures: ArrayBufferSlice[]): AnimatedTexture[] {
     return bindings.map((binding) => ({
@@ -105,39 +51,6 @@ function resolveAnimatedMaterialTextures(bindings: readonly AnimatedMaterialText
         frameDuration: binding.frameDuration,
         frames: binding.textureIDs.map((textureID) => textures[textureID]),
     }));
-}
-
-function createSpriteVertexBuffer(sprite: SpriteData, quadCount = 1): ArrayBufferSlice {
-    const buffer = new ArrayBuffer(quadCount * 4 * 0x10);
-    const view = new DataView(buffer);
-    const textureCoordinates = [
-        0, 0,
-        sprite.width << 5, 0,
-        sprite.width << 5, sprite.height << 5,
-        0, sprite.height << 5,
-    ];
-    for (let quad = 0; quad < quadCount; quad++) {
-        for (let i = 0; i < 4; i++) {
-            const offs = (quad * 4 + i) * 0x10;
-            view.setInt16(offs + 0x08, textureCoordinates[i * 2]);
-            view.setInt16(offs + 0x0A, textureCoordinates[i * 2 + 1]);
-            view.setUint8(offs + 0x0C, 0xFF);
-            view.setUint8(offs + 0x0D, 0xFF);
-            view.setUint8(offs + 0x0E, 0xFF);
-            view.setUint8(offs + 0x0F, 0xFF);
-        }
-    }
-    return new ArrayBufferSlice(buffer);
-}
-
-function generatedSurfaceHeight(surface: GeneratedSurface, x: number, z: number, tick: number): number {
-    const phaseS = tick * surface.phaseSpeedS;
-    const phaseT = tick * surface.phaseSpeedT;
-    const angleS = (phaseS + Math.trunc(surface.frequencyS * x)) % 0x0FFF;
-    const angleT = (phaseT + Math.trunc(surface.frequencyT * z)) % 0x0FFF;
-    return surface.baseY
-        + Math.sin(angleS * Math.PI * 2 / 0x1000) * surface.amplitudeS
-        + Math.sin(angleT * Math.PI * 2 / 0x1000) * surface.amplitudeT;
 }
 
 function createGeneratedSurfaceVertexBuffer(surface: GeneratedSurface): ArrayBufferSlice {
@@ -167,802 +80,6 @@ function createGeneratedSurfaceVertexBuffer(surface: GeneratedSurface): ArrayBuf
     }
     return new ArrayBufferSlice(buffer);
 }
-
-interface FogParams {
-    near: number;
-    far: number;
-    color: readonly [number, number, number, number];
-}
-
-function fogPositionToViewDistance(position: number, clipNear: number, clipFar: number): number {
-    const normalizedPosition = position / 1000;
-    return clipNear * clipFar / (clipFar - normalizedPosition * (clipFar - clipNear));
-}
-
-class DrawCallInstance {
-    private textureEntry: Texture[] = [];
-    private animatedTextureMappings: TextureMapping[][] = [];
-    private vertexColorsEnabled = true;
-    private texturesEnabled = true;
-    private monochromeVertexColorsEnabled = false;
-    private alphaVisualizerEnabled = false;
-    private fogEnabled = false;
-    private megaStateFlags: Partial<GfxMegaStateDescriptor>;
-    private program!: DeviceProgram;
-    private gfxProgram: GfxProgram | null = null;
-    private textureMappings = nArray(2, () => new TextureMapping());
-    private isTranslucent = false;
-    private crossfadeDuration = 0;
-    private viewMatrix = mat4.create();
-    private boneModelViewMatrix = mat4.create();
-    private textureMatrix = mat4.create();
-    public visible = true;
-
-    constructor(device: GfxDevice, cache: GfxRenderCache, gpuTextureCache: GPUTextureCache, sharedOutput: RSPSharedOutput, private drawCall: DrawCall, private firstIndex: number, private fogParams: FogParams, private boneMatrices: mat4[] | undefined) {
-        const linearFiltering = ((drawCall.DP_OtherModeH >>> OtherModeH_Layout.G_MDSFT_TEXTFILT) & 0x03) === TextFilt.G_TF_BILERP;
-        for (let i = 0; i < this.textureMappings.length; i++) {
-            const textureIndex = drawCall.textureIndices[i];
-            if (textureIndex === undefined)
-                continue;
-            const tex = sharedOutput.textureCache.textures[textureIndex];
-
-            if (tex) {
-                this.textureEntry[i] = tex;
-                this.textureMappings[i].gfxTexture = gpuTextureCache.getTexture(device, tex);
-                this.textureMappings[i].gfxSampler = translateSampler(cache, tex, linearFiltering);
-            }
-
-            const animation = drawCall.textureAnimations[i];
-            if (animation !== undefined) {
-                const entries = animation.textureIndices.map((index) => sharedOutput.textureCache.textures[index]);
-                this.animatedTextureMappings[i] = entries.map((entry, frame) => {
-                    if (frame === 0)
-                        return this.textureMappings[i];
-                    const mapping = new TextureMapping();
-                    mapping.gfxTexture = gpuTextureCache.getTexture(device, entry);
-                    mapping.gfxSampler = translateSampler(cache, entry, linearFiltering);
-                    return mapping;
-                });
-            }
-        }
-        const crossfade0 = drawCall.textureAnimations[0]?.crossfadeGroup;
-        const crossfade1 = drawCall.textureAnimations[1]?.crossfadeGroup;
-        if (crossfade0 !== null && crossfade0 !== undefined && crossfade0 === crossfade1)
-            this.crossfadeDuration = Math.max(drawCall.textureAnimations[0]!.frameDuration, 1);
-
-        this.megaStateFlags = translateBlendMode(this.drawCall.SP_GeometryMode, this.drawCall.DP_OtherModeL);
-        this.isTranslucent = this.crossfadeDuration > 0 || renderModeIsTranslucent(this.megaStateFlags);
-        this.setBackfaceCullingEnabled(true);
-        this.createProgram();
-    }
-
-    private createProgram(): void {
-        const program = new F3DEX_Program(this.drawCall.DP_OtherModeH, this.drawCall.DP_OtherModeL, this.drawCall.DP_Combine);
-        program.defines.set('BONE_MATRIX_COUNT', (this.boneMatrices?.length ?? 1).toString());
-
-        if (this.texturesEnabled && this.textureEntry.length)
-            program.defines.set('USE_TEXTURE', '1');
-
-        if (!!(this.drawCall.SP_GeometryMode & RSP_Geometry.G_LIGHTING))
-            program.defines.set('LIGHTING', '1');
-
-        if (this.vertexColorsEnabled && (this.drawCall.SP_GeometryMode & RSP_Geometry.G_SHADE) !== 0)
-            program.defines.set('USE_VERTEX_COLOR', '1');
-
-        if (this.drawCall.SP_GeometryMode & RSP_Geometry.G_TEXTURE_GEN)
-            program.defines.set('TEXTURE_GEN', '1');
-
-        // many display lists seem to set this flag without setting texture_gen,
-        // despite this one being dependent on it
-        if (this.drawCall.SP_GeometryMode & RSP_Geometry.G_TEXTURE_GEN_LINEAR)
-            program.defines.set('TEXTURE_GEN_LINEAR', '1');
-
-        if (this.fogEnabled && (this.drawCall.SP_GeometryMode & RSP_Geometry.G_FOG)) {
-            program.defines.set('USE_FOG', '1');
-        }
-
-        if (this.monochromeVertexColorsEnabled)
-            program.defines.set('USE_MONOCHROME_VERTEX_COLOR', '1');
-
-        if (this.alphaVisualizerEnabled)
-            program.defines.set('USE_ALPHA_VISUALIZER', '1');
-
-        if (this.crossfadeDuration > 0)
-            program.defines.set('EXTRA_COMBINE', '1');
-
-        this.program = program;
-        this.gfxProgram = null;
-    }
-
-    public setBackfaceCullingEnabled(v: boolean): void {
-        const cullMode = v ? translateCullMode(this.drawCall.SP_GeometryMode) : GfxCullMode.None;
-        this.megaStateFlags.cullMode = cullMode;
-    }
-
-    public setVertexColorsEnabled(v: boolean): void {
-        this.vertexColorsEnabled = v;
-        this.createProgram();
-    }
-
-    public setTexturesEnabled(v: boolean): void {
-        this.texturesEnabled = v;
-        this.createProgram();
-    }
-
-    public setFogEnabled(v: boolean): void {
-        this.fogEnabled = v;
-        this.createProgram();
-    }
-
-    public setMonochromeVertexColorsEnabled(v: boolean): void {
-        this.monochromeVertexColorsEnabled = v;
-        this.createProgram();
-    }
-
-    public setAlphaVisualizerEnabled(v: boolean): void {
-        this.alphaVisualizerEnabled = v;
-        this.createProgram();
-    }
-
-    private computeTextureMatrix(m: mat4, textureEntryIndex: number, time: number): void {
-        if (this.textureEntry[textureEntryIndex] !== undefined) {
-            const entry = this.textureEntry[textureEntryIndex];
-            calcTextureMatrixFromRSPState(m, this.drawCall.SP_TextureState.s, this.drawCall.SP_TextureState.t, entry.width, entry.height, entry.tile.shifts, entry.tile.shiftt);
-            const speed = this.drawCall.textureScrollSpeeds[textureEntryIndex] ?? 0;
-            if (speed !== 0) {
-                const ticks = Math.floor(time / (1000 / 30));
-                if (ticks > 0) {
-                    const cycle = (Math.floor(255 / speed) + 1) * speed;
-                    const tileOffset = 255 - (((ticks - 1) * speed) % cycle);
-                    m[13] -= (tileOffset / 4) / entry.height;
-                }
-            }
-        } else {
-            mat4.identity(m);
-        }
-    }
-
-    public prepareToRender(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput, modelMatrix: mat4, isSkybox: boolean, primAlphaMultiplier = 1, primColorMultiplier: vec3 | null = null, sprites: readonly SpriteBillboard[] | null = null): void {
-        if (!this.visible)
-            return;
-
-        if (this.gfxProgram === null)
-            this.gfxProgram = renderInstManager.gfxRenderCache.createProgram(this.program);
-
-        const animationTick = viewerInput.time / (1000 / 30);
-        for (let i = 0; i < this.animatedTextureMappings.length; i++) {
-            const mappings = this.animatedTextureMappings[i];
-            if (mappings === undefined)
-                continue;
-            const animation = this.drawCall.textureAnimations[i]!;
-            const frameDuration = Math.max(animation.frameDuration, 1);
-            const frameOffset = animation.frameOffset;
-            const frame = (Math.floor(animationTick / frameDuration) + frameOffset) % mappings.length;
-            this.textureMappings[i] = mappings[frame];
-        }
-        if (sprites !== null) {
-            for (let i = 0; i < sprites.length; i++) {
-                if (sprites[i].fade > 0)
-                    this.prepareSingleRenderInst(renderInstManager, viewerInput, modelMatrix, isSkybox, primAlphaMultiplier * sprites[i].fade, primColorMultiplier, 6, this.firstIndex + i * 6);
-            }
-        } else {
-            this.prepareSingleRenderInst(renderInstManager, viewerInput, modelMatrix, isSkybox, primAlphaMultiplier, primColorMultiplier, this.drawCall.indexCount, this.firstIndex);
-        }
-    }
-
-    private prepareSingleRenderInst(renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput, modelMatrix: mat4, isSkybox: boolean, primAlphaMultiplier: number, primColorMultiplier: vec3 | null, indexCount: number, firstIndex: number): void {
-        const renderInst = renderInstManager.newRenderInst();
-        if (this.isTranslucent)
-            renderInst.sortKey = makeSortKey(GfxRendererLayer.TRANSLUCENT);
-        renderInst.setGfxProgram(this.gfxProgram!);
-        renderInst.setSamplerBindingsFromTextureMappings(this.textureMappings);
-        renderInst.setMegaStateFlags(this.megaStateFlags);
-        renderInst.setDrawCount(indexCount, firstIndex);
-
-        const usesFog = this.fogEnabled && (this.drawCall.SP_GeometryMode & RSP_Geometry.G_FOG) !== 0;
-        const boneMatrixCount = this.boneMatrices?.length ?? 1;
-        let offs = renderInst.allocateUniformBuffer(F3DEX_Program.ub_DrawParams, 12 * boneMatrixCount + 8*2 + (usesFog ? 8 : 0));
-        const mappedF32 = renderInst.mapUniformBufferF32(F3DEX_Program.ub_DrawParams);
-
-        if (isSkybox)
-            computeViewMatrixSkybox(this.viewMatrix, viewerInput.camera);
-        else
-            computeViewMatrix(this.viewMatrix, viewerInput.camera);
-        mat4.mul(this.viewMatrix, this.viewMatrix, modelMatrix);
-
-        if (this.boneMatrices !== undefined) {
-            for (const boneMatrix of this.boneMatrices) {
-                mat4.mul(this.boneModelViewMatrix, this.viewMatrix, boneMatrix);
-                offs += fillMatrix4x3(mappedF32, offs, this.boneModelViewMatrix); // u_BoneMatrix
-            }
-        } else {
-            offs += fillMatrix4x3(mappedF32, offs, this.viewMatrix); // u_BoneMatrix
-        }
-
-        this.computeTextureMatrix(this.textureMatrix, 0, viewerInput.time);
-        offs += fillMatrix4x2(mappedF32, offs, this.textureMatrix); // u_TexMatrix[0]
-
-        this.computeTextureMatrix(this.textureMatrix, 1, viewerInput.time);
-        offs += fillMatrix4x2(mappedF32, offs, this.textureMatrix); // u_TexMatrix[1]
-
-        if (usesFog) {
-            offs += fillVec4(mappedF32, offs, this.fogParams.near, this.fogParams.far, 0, 0);
-            const fogColor = this.fogParams.color;
-            offs += fillVec4(mappedF32, offs, fogColor[0], fogColor[1], fogColor[2], fogColor[3]);
-        }
-
-        offs = renderInst.allocateUniformBuffer(F3DEX_Program.ub_CombineParams, this.crossfadeDuration > 0 ? 12 : 8);
-        const comb = renderInst.mapUniformBufferF32(F3DEX_Program.ub_CombineParams);
-        const primColor = this.drawCall.DP_PrimColor;
-        offs += fillVec4(comb, offs,
-            primColor[0] * (primColorMultiplier?.[0] ?? 1),
-            primColor[1] * (primColorMultiplier?.[1] ?? 1),
-            primColor[2] * (primColorMultiplier?.[2] ?? 1),
-            primColor[3] * primAlphaMultiplier);
-        const envColor = this.drawCall.DP_EnvColor;
-        offs += fillVec4(comb, offs, envColor[0], envColor[1], envColor[2], envColor[3]);
-        if (this.crossfadeDuration > 0) {
-            // Interpolate the 30Hz game tick in PRIM_LOD_FRAC for smoother crossfades.
-            const animationTick = viewerInput.time / (1000 / 30);
-            const blend = (animationTick % this.crossfadeDuration) / this.crossfadeDuration;
-            offs += fillVec4(comb, offs, blend, 0, 0, 0);
-        }
-        renderInstManager.submitRenderInst(renderInst);
-    }
-
-    public destroy(device: GfxDevice): void {
-    }
-}
-
-function makeVertexBufferData(v: Vertex[], actorAnimation: ActorAnimationState | undefined): Float32Array {
-    const buf = new Float32Array(10 * v.length);
-    let j = 0;
-    for (let i = 0; i < v.length; i++) {
-        buf[j++] = v[i].x;
-        buf[j++] = v[i].y;
-        buf[j++] = v[i].z;
-        buf[j++] = actorAnimation !== undefined
-            ? Math.min(v[i].matrixIndex, actorAnimation.pose.boneMatrices.length - 1)
-            : 1.0;
-
-        buf[j++] = v[i].tx;
-        buf[j++] = v[i].ty;
-
-        buf[j++] = v[i].c0;
-        buf[j++] = v[i].c1;
-        buf[j++] = v[i].c2;
-        buf[j++] = v[i].a;
-    }
-    return buf;
-}
-
-export class RenderData {
-    public vertexBuffer: GfxBuffer;
-    public inputLayout: GfxInputLayout;
-    public vertexBufferDescriptors: GfxVertexBufferDescriptor[];
-    public indexBufferDescriptor: GfxIndexBufferDescriptor;
-    public vertexBufferData: Float32Array;
-    public indexBuffer: GfxBuffer;
-    public vertexStart: number;
-    public indexStart: number;
-
-    constructor(device: GfxDevice, cache: GfxRenderCache, mesh: Mesh, dynamic = false) {
-        const sharedOutput = mesh.sharedOutput;
-        const drawCalls = mesh.rspOutput?.drawCalls ?? [];
-        this.indexStart = drawCalls.reduce(
-            (start, drawCall) => Math.min(start, drawCall.firstIndex),
-            sharedOutput.indices.length,
-        );
-        const indexEnd = drawCalls.reduce(
-            (end, drawCall) => Math.max(end, drawCall.firstIndex + drawCall.indexCount),
-            this.indexStart,
-        );
-        const sharedIndices = sharedOutput.indices.slice(this.indexStart, indexEnd);
-        let vertexStart = sharedIndices.reduce(
-            (start, vertexIndex) => Math.min(start, vertexIndex),
-            sharedOutput.vertices.length,
-        );
-        let vertexEnd = sharedIndices.reduce(
-            (end, vertexIndex) => Math.max(end, vertexIndex + 1),
-            vertexStart,
-        );
-        const dynamicVertexRange = getDynamicVertexRange(mesh);
-        if (dynamicVertexRange !== null) {
-            vertexStart = Math.min(vertexStart, dynamicVertexRange.start);
-            vertexEnd = Math.max(vertexEnd, dynamicVertexRange.end);
-        }
-        this.vertexStart = vertexStart;
-
-        assert(vertexEnd - this.vertexStart <= 0xFFFFFFFF);
-        this.vertexBufferData = makeVertexBufferData(sharedOutput.vertices.slice(this.vertexStart, vertexEnd), mesh.actorAnimation);
-        this.vertexBuffer = createBufferFromData(device, GfxBufferUsage.Vertex, dynamic ? GfxBufferFrequencyHint.Dynamic : GfxBufferFrequencyHint.Static, this.vertexBufferData.buffer);
-
-        const indexBufferData = Uint32Array.from(sharedIndices, (vertexIndex) => vertexIndex - this.vertexStart);
-        this.indexBuffer = createBufferFromData(device, GfxBufferUsage.Index, GfxBufferFrequencyHint.Static, indexBufferData.buffer);
-
-        const vertexAttributeDescriptors: GfxVertexAttributeDescriptor[] = [
-            { location: F3DEX_Program.a_Position, bufferIndex: 0, format: GfxFormat.F32_RGBA, bufferByteOffset: 0*0x04, },
-            { location: F3DEX_Program.a_TexCoord, bufferIndex: 0, format: GfxFormat.F32_RG,   bufferByteOffset: 4*0x04, },
-            { location: F3DEX_Program.a_Color   , bufferIndex: 0, format: GfxFormat.F32_RGBA, bufferByteOffset: 6*0x04, },
-        ];
-
-        const vertexBufferDescriptors: GfxInputLayoutBufferDescriptor[] = [
-            { byteStride: 10*0x04, frequency: GfxVertexBufferFrequency.PerVertex, },
-        ];
-
-        this.inputLayout = cache.createInputLayout({
-            indexBufferFormat: GfxFormat.U32_R,
-            vertexBufferDescriptors,
-            vertexAttributeDescriptors,
-        });
-
-        this.vertexBufferDescriptors = [{ buffer: this.vertexBuffer }];
-        this.indexBufferDescriptor = { buffer: this.indexBuffer };
-    }
-
-    public destroy(device: GfxDevice): void {
-        device.destroyBuffer(this.indexBuffer);
-        device.destroyBuffer(this.vertexBuffer);
-    }
-}
-
-class SpriteBillboard {
-    private static readonly signs = [-1, 1, 1, 1, 1, -1, -1, -1];
-    public fade = 1;
-    private spawnTick: number | undefined;
-    private lifetime: number | undefined;
-    private loopTicks: number | undefined;
-    private velocityY: number | undefined;
-    public maxDistance: number | undefined;
-    public fadeStartDistance: number | undefined;
-
-    constructor(
-        public firstVertex: number,
-        public origin: vec3,
-        private halfWidth: number,
-        private halfHeight: number,
-        options: {
-            spawnTick?: number;
-            lifetime?: number;
-            loopTicks?: number;
-            velocityY?: number;
-            maxDistance?: number;
-            fadeStartDistance?: number;
-        },
-    ) {
-        this.spawnTick = options.spawnTick;
-        this.lifetime = options.lifetime;
-        this.loopTicks = options.loopTicks;
-        this.velocityY = options.velocityY;
-        this.maxDistance = options.maxDistance;
-        this.fadeStartDistance = options.fadeStartDistance;
-    }
-
-    public update(vertexBufferData: Float32Array, vertexBufferFirstVertex: number, cameraMatrix: mat4, tick: number): void {
-        const age = this.spawnTick === undefined
-            ? 0
-            : ((tick - this.spawnTick + this.loopTicks!) % this.loopTicks!);
-        const dx = cameraMatrix[12] - this.origin[0];
-        const dy = cameraMatrix[13] - this.origin[1];
-        const dz = cameraMatrix[14] - this.origin[2];
-        const distance = Math.sqrt(dx * dx + dy * dy + dz * dz);
-        const withinDistance = this.maxDistance === undefined || distance <= this.maxDistance;
-        const active = withinDistance && (this.spawnTick === undefined || age < this.lifetime!);
-        this.fade = active ? 1 : 0;
-        if (this.fadeStartDistance !== undefined) {
-            this.fade = !active ? 0 : distance < this.fadeStartDistance
-                ? 1
-                : Math.max(0, Math.min(1, (this.maxDistance! - distance) / (this.maxDistance! - this.fadeStartDistance)));
-        }
-        const centerX = this.origin[0];
-        const centerY = this.origin[1] + age * (this.velocityY ?? 0);
-        const centerZ = this.origin[2];
-        for (let i = 0; i < 4; i++) {
-            const vertex = (this.firstVertex + i - vertexBufferFirstVertex) * 10;
-            if (!active) {
-                vertexBufferData[vertex + 0] = this.origin[0];
-                vertexBufferData[vertex + 1] = this.origin[1];
-                vertexBufferData[vertex + 2] = this.origin[2];
-                continue;
-            }
-            const rightOffset = this.halfWidth * SpriteBillboard.signs[i * 2];
-            const upOffset = this.halfHeight * SpriteBillboard.signs[i * 2 + 1];
-            vertexBufferData[vertex + 0] = centerX
-                + cameraMatrix[0] * rightOffset
-                + cameraMatrix[4] * upOffset;
-            vertexBufferData[vertex + 1] = centerY
-                + cameraMatrix[1] * rightOffset
-                + cameraMatrix[5] * upOffset;
-            vertexBufferData[vertex + 2] = centerZ
-                + cameraMatrix[2] * rightOffset
-                + cameraMatrix[6] * upOffset;
-        }
-    }
-}
-
-export interface Mesh {
-    sharedOutput: RSPSharedOutput;
-    rspState: RSPState;
-    rspOutput: RSPOutput | null;
-    generatedSurfaceAnimation?: {
-        surface: GeneratedSurface;
-        firstVertex: number;
-        vertexCount: number;
-    };
-    dynamicLighting?: DynamicLighting;
-    actorAnimation?: ActorAnimationState;
-    propAnimation?: PropAnimationState;
-    spriteBillboards?: SpriteBillboard[];
-}
-
-function getDynamicVertexRange(mesh: Mesh): { start: number, end: number } | null {
-    let rangeStart = Infinity, rangeEnd = -Infinity;
-    const include = (start: number, count: number): void => {
-        rangeStart = Math.min(rangeStart, start);
-        rangeEnd = Math.max(rangeEnd, start + count);
-    };
-    if (mesh.generatedSurfaceAnimation !== undefined)
-        include(mesh.generatedSurfaceAnimation.firstVertex, mesh.generatedSurfaceAnimation.vertexCount);
-    if (mesh.propAnimation !== undefined)
-        for (const vertexOffset of mesh.propAnimation.vertexOffsets)
-            include(mesh.propAnimation.firstVertex + vertexOffset, 1);
-    for (const sprite of mesh.spriteBillboards ?? [])
-        include(sprite.firstVertex, 4);
-    for (const vertexIndex of mesh.dynamicLighting?.vertexIndices ?? [])
-        include(vertexIndex, 1);
-    return rangeStart <= rangeEnd ? { start: rangeStart, end: rangeEnd } : null;
-}
-
-export class MeshData {
-    public renderData: RenderData;
-    public cullBounds: MeshBounds;
-    private lightingDirty: boolean;
-    private dirtyVertexRange: { start: number, end: number } | null = null;
-    public dynamicLightingEnabled = true;
-
-    constructor(device: GfxDevice, cache: GfxRenderCache, public mesh: Mesh) {
-        this.renderData = new RenderData(device, cache, mesh, mesh.generatedSurfaceAnimation !== undefined || mesh.spriteBillboards !== undefined || mesh.dynamicLighting !== undefined || mesh.propAnimation !== undefined);
-        this.lightingDirty = mesh.dynamicLighting !== undefined;
-        this.cullBounds = new MeshBounds(
-            mesh.sharedOutput,
-            mesh.rspOutput,
-            mesh.actorAnimation?.boundingBox,
-            mesh.propAnimation?.translationBounds,
-            mesh.propAnimation !== undefined ? 0.5 : 0,
-        );
-
-        const dynamicVertexRange = getDynamicVertexRange(mesh);
-        if (dynamicVertexRange !== null) {
-            this.dirtyVertexRange = {
-                start: dynamicVertexRange.start - this.renderData.vertexStart,
-                end: dynamicVertexRange.end - this.renderData.vertexStart,
-            };
-        }
-    }
-
-    public setDynamicLightingEnabled(enabled: boolean): void {
-        this.dynamicLightingEnabled = enabled;
-        this.lightingDirty = true;
-    }
-
-    public update(device: GfxDevice, viewerInput: Viewer.ViewerRenderInput, activeLightCache: ActiveLightCache): void {
-        const animation = this.mesh.generatedSurfaceAnimation;
-        const sprites = this.mesh.spriteBillboards;
-        const lighting = this.mesh.dynamicLighting;
-        const actorAnimation = this.mesh.actorAnimation;
-        const propAnimation = this.mesh.propAnimation;
-        const tick = Math.floor(viewerInput.time / (1000 / 30));
-        const lightingIsDynamic = lighting !== undefined && lighting.lights.length > 0 && this.dynamicLightingEnabled;
-
-        if (animation !== undefined) {
-            const surface = animation.surface;
-            const amplitude = surface.amplitudeS + surface.amplitudeT;
-            for (let i = 0; i < animation.vertexCount; i++) {
-                const vertexIndex = animation.firstVertex + i;
-                const vertex = this.mesh.sharedOutput.vertices[vertexIndex];
-                const y = generatedSurfaceHeight(surface, vertex.x / 3, vertex.z / 3, tick);
-                const alpha = Math.max(0, Math.min(0xFF, Math.trunc(
-                    ((y - surface.baseY) / amplitude) * surface.alphaRange + surface.alphaBase,
-                )));
-                const localVertex = (vertexIndex - this.renderData.vertexStart) * 10;
-                this.renderData.vertexBufferData[localVertex + 1] = Math.trunc(y * 3);
-                this.renderData.vertexBufferData[localVertex + 9] = alpha / 0xFF;
-            }
-        }
-
-        if (lighting !== undefined && (lightingIsDynamic || this.lightingDirty)) {
-            updateDynamicLighting(lighting, this.mesh.sharedOutput.vertices, this.renderData.vertexBufferData, this.renderData.vertexStart, activeLightCache, this.dynamicLightingEnabled);
-            this.lightingDirty = false;
-        }
-        if (actorAnimation !== undefined)
-            actorAnimation.pose.update(tick);
-        if (propAnimation !== undefined)
-            propAnimation.update(this.renderData.vertexBufferData, this.renderData.vertexStart, tick);
-
-        for (const sprite of sprites ?? [])
-            sprite.update(this.renderData.vertexBufferData, this.renderData.vertexStart, viewerInput.camera.worldMatrix, tick);
-        if (this.dirtyVertexRange !== null) {
-            const byteOffset = this.dirtyVertexRange.start * 10 * 4;
-            const byteLength = (this.dirtyVertexRange.end - this.dirtyVertexRange.start) * 10 * 4;
-            device.uploadBufferData(
-                this.renderData.vertexBuffer,
-                byteOffset,
-                new Uint8Array(this.renderData.vertexBufferData.buffer, byteOffset, byteLength),
-            );
-        }
-    }
-
-    public destroy(device: GfxDevice): void {
-        this.renderData.destroy(device);
-    }
-}
-
-enum SceneRenderLayer {
-    MapGeometry,
-    Props,
-    Actors,
-    Surfaces,
-    Effects,
-}
-
-class MeshRenderer {
-    public drawCallInstances: DrawCallInstance[] = [];
-
-    public prepareToRender(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput, modelMatrix: mat4, isSkybox: boolean, primAlphaMultiplier = 1, primColorMultiplier: vec3 | null = null, sprites: readonly SpriteBillboard[] | null = null): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].prepareToRender(device, renderInstManager, viewerInput, modelMatrix, isSkybox, primAlphaMultiplier, primColorMultiplier, sprites);
-    }
-
-    public setBackfaceCullingEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setBackfaceCullingEnabled(v);
-    }
-
-    public setVertexColorsEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setVertexColorsEnabled(v);
-    }
-
-    public setTexturesEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setTexturesEnabled(v);
-    }
-
-    public setFogEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setFogEnabled(v);
-    }
-
-    public setMonochromeVertexColorsEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setMonochromeVertexColorsEnabled(v);
-    }
-
-    public setAlphaVisualizerEnabled(v: boolean): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].setAlphaVisualizerEnabled(v);
-    }
-
-    public destroy(device: GfxDevice): void {
-        for (let i = 0; i < this.drawCallInstances.length; i++)
-            this.drawCallInstances[i].destroy(device);
-    }
-}
-
-export class RootMeshRenderer {
-    private visible = true;
-    private cullBoundingBox: AABB | null = null;
-    private megaStateFlags: Partial<GfxMegaStateDescriptor>;
-    public isSkybox = false;
-    public sortKeyBase = makeSortKey(GfxRendererLayer.OPAQUE);
-    public modelMatrix = mat4.create();
-    public distanceFade: { origin: vec3; startDistance: number; endDistance: number } | null = null;
-    private rootTransformAnimation: {
-        baseMatrix: mat4;
-        rotationYRadiansPerTick: number;
-        positionYAmplitude: number;
-        positionYRadiansPerTick: number;
-    } | null = null;
-    private computeLookAt = false;
-    private cameraBillboard: { origin: vec3; scale: number } | null = null;
-    private objectLighting: ObjectLighting | null = null;
-    private objectLightColor = vec3.create();
-    private lookAtPosition = vec3.create();
-    private lookAtMatrix = mat4.create();
-
-    public objectFlags = 0;
-    private rootNodeRenderer: MeshRenderer;
-    private ownsRootNodeRenderer: boolean;
-
-    constructor(
-        device: GfxDevice,
-        cache: GfxRenderCache,
-        private geometryData: MeshData,
-        public renderLayer: SceneRenderLayer,
-        private fogParams: FogParams,
-        private gpuTextureCache: GPUTextureCache,
-        sharedRenderer: RootMeshRenderer | null = null,
-    ) {
-        this.megaStateFlags = {};
-        setAttachmentStateSimple(this.megaStateFlags, {
-            blendMode: GfxBlendMode.Add,
-            blendSrcFactor: GfxBlendFactor.SrcAlpha,
-            blendDstFactor: GfxBlendFactor.OneMinusSrcAlpha,
-        });
-
-        const geo = this.geometryData.mesh;
-        this.computeLookAt = geo.rspOutput?.drawCalls.some((drawCall) => {
-            const requiredModes = RSP_Geometry.G_LIGHTING | RSP_Geometry.G_TEXTURE_GEN;
-            return (drawCall.SP_GeometryMode & requiredModes) === requiredModes;
-        }) ?? false;
-
-        if (sharedRenderer !== null) {
-            assert(sharedRenderer.geometryData === geometryData);
-            this.rootNodeRenderer = sharedRenderer.rootNodeRenderer;
-            this.ownsRootNodeRenderer = false;
-        } else {
-            this.rootNodeRenderer = this.buildGeoNodeRenderer(device, cache, geo);
-            this.ownsRootNodeRenderer = true;
-        }
-    }
-
-    private buildGeoNodeRenderer(device: GfxDevice, cache: GfxRenderCache, node: Mesh): MeshRenderer {
-        const geoNodeRenderer = new MeshRenderer();
-
-        if (node.rspOutput !== null) {
-            for (let i = 0; i < node.rspOutput.drawCalls.length; i++) {
-                const drawCall = node.rspOutput.drawCalls[i];
-                const drawCallInstance = new DrawCallInstance(device, cache, this.gpuTextureCache, node.sharedOutput, drawCall, drawCall.firstIndex - this.geometryData.renderData.indexStart, this.fogParams, node.actorAnimation?.pose.boneMatrices);
-                geoNodeRenderer.drawCallInstances.push(drawCallInstance);
-            }
-        }
-
-        return geoNodeRenderer;
-    }
-
-    public setBackfaceCullingEnabled(v: boolean): void {
-        this.rootNodeRenderer.setBackfaceCullingEnabled(v);
-    }
-
-    public setVertexColorsEnabled(v: boolean): void {
-        this.rootNodeRenderer.setVertexColorsEnabled(v);
-    }
-
-    public setTexturesEnabled(v: boolean): void {
-        this.rootNodeRenderer.setTexturesEnabled(v);
-    }
-
-    public setFogEnabled(v: boolean): void {
-        this.rootNodeRenderer.setFogEnabled(v);
-    }
-
-    public setMonochromeVertexColorsEnabled(v: boolean): void {
-        this.rootNodeRenderer.setMonochromeVertexColorsEnabled(v);
-    }
-
-    public setAlphaVisualizerEnabled(v: boolean): void {
-        this.rootNodeRenderer.setAlphaVisualizerEnabled(v);
-    }
-
-    public setVisible(v: boolean): void {
-        this.visible = v;
-    }
-
-    public setCullBoundingBox(boundingBox: AABB | null): void {
-        this.cullBoundingBox = boundingBox;
-    }
-
-    public computeWorldBoundingBox(): AABB | null {
-        return this.geometryData.cullBounds.computeWorld(this.modelMatrix, this.rootTransformAnimation);
-    }
-
-    public setRotationYAnimation(anglePerTick: number): void {
-        this.ensureRootTransformAnimation();
-        this.rootTransformAnimation!.rotationYRadiansPerTick = anglePerTick / 0x1000 * Math.PI * 2;
-    }
-
-    public setPositionYAnimation(amplitude: number, anglePerTick: number): void {
-        this.ensureRootTransformAnimation();
-        this.rootTransformAnimation!.positionYAmplitude = amplitude;
-        this.rootTransformAnimation!.positionYRadiansPerTick = anglePerTick / 0x1000 * Math.PI * 2;
-    }
-
-    public setCameraBillboard(origin: vec3, scale: number): void {
-        this.cameraBillboard = { origin, scale };
-    }
-
-    private ensureRootTransformAnimation(): void {
-        if (this.rootTransformAnimation !== null)
-            return;
-        this.rootTransformAnimation = {
-            baseMatrix: mat4.clone(this.modelMatrix),
-            rotationYRadiansPerTick: 0,
-            positionYAmplitude: 0,
-            positionYRadiansPerTick: 0,
-        };
-    }
-
-    public setObjectLighting(lighting: ObjectLighting): void {
-        this.objectLighting = lighting;
-    }
-
-    public prepareToRender(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput, activeLightCache: ActiveLightCache): void {
-        if (!this.visible)
-            return;
-        if (this.cullBoundingBox !== null && !viewerInput.camera.frustum.contains(this.cullBoundingBox))
-            return;
-
-        if (this.rootTransformAnimation !== null) {
-            const animation = this.rootTransformAnimation;
-            mat4.copy(this.modelMatrix, animation.baseMatrix);
-            const tick = Math.floor(viewerInput.time / (1000 / 30));
-            mat4.rotateY(this.modelMatrix, this.modelMatrix, tick * animation.rotationYRadiansPerTick);
-            this.modelMatrix[13] += Math.sin(tick * animation.positionYRadiansPerTick)
-                * animation.positionYAmplitude;
-        } else if (this.cameraBillboard !== null) {
-            const billboard = this.cameraBillboard;
-            scaleMatrix(this.modelMatrix, viewerInput.camera.worldMatrix, billboard.scale);
-            setMatrixTranslation(this.modelMatrix, billboard.origin);
-        }
-
-        let primAlphaMultiplier = 1;
-        if (this.distanceFade !== null) {
-            const camera = viewerInput.camera.worldMatrix;
-            const dx = camera[12] - this.distanceFade.origin[0];
-            const dy = camera[13] - this.distanceFade.origin[1];
-            const dz = camera[14] - this.distanceFade.origin[2];
-            const distance = Math.sqrt(dx * dx + dy * dy + dz * dz);
-            if (distance >= this.distanceFade.endDistance)
-                return;
-            if (distance > this.distanceFade.startDistance) {
-                primAlphaMultiplier = (this.distanceFade.endDistance - distance)
-                    / (this.distanceFade.endDistance - this.distanceFade.startDistance);
-            }
-        }
-
-        this.geometryData.update(device, viewerInput, activeLightCache);
-        const renderData = this.geometryData.renderData;
-
-        const template = renderInstManager.pushTemplate();
-        template.setBindingLayouts(bindingLayouts);
-        template.setVertexInput(renderData.inputLayout, renderData.vertexBufferDescriptors, renderData.indexBufferDescriptor);
-        template.setMegaStateFlags(this.megaStateFlags);
-
-        template.sortKey = this.sortKeyBase;
-
-        const sceneParamsSize = 16 + (this.computeLookAt ? 8 : 0);
-
-        let offs = template.allocateUniformBuffer(F3DEX_Program.ub_SceneParams, sceneParamsSize);
-        const mappedF32 = template.mapUniformBufferF32(F3DEX_Program.ub_SceneParams);
-        offs += fillMatrix4x4(mappedF32, offs, viewerInput.camera.projectionMatrix);
-
-        if (this.computeLookAt) {
-            // compute lookat X and Y in view space, since that's the transform the shader will have
-            mat4.getTranslation(this.lookAtPosition, this.modelMatrix);
-            vec3.transformMat4(this.lookAtPosition, this.lookAtPosition, viewerInput.camera.viewMatrix);
-
-            mat4.lookAt(this.lookAtMatrix, Vec3Zero, this.lookAtPosition, Vec3UnitY);
-            offs += fillVec4(mappedF32, offs, this.lookAtMatrix[0], this.lookAtMatrix[4], this.lookAtMatrix[8]);
-            offs += fillVec4(mappedF32, offs, this.lookAtMatrix[1], this.lookAtMatrix[5], this.lookAtMatrix[9]);
-        }
-
-        const objectLightColor = this.objectLighting !== null
-            ? sampleObjectLighting(this.objectLightColor, this.objectLighting, activeLightCache, this.geometryData.dynamicLightingEnabled)
-            : null;
-        this.rootNodeRenderer.prepareToRender(device, renderInstManager, viewerInput, this.modelMatrix, this.isSkybox, primAlphaMultiplier, objectLightColor, this.geometryData.mesh.spriteBillboards ?? null);
-
-        renderInstManager.popTemplate();
-    }
-
-    public destroy(device: GfxDevice): void {
-        if (this.ownsRootNodeRenderer)
-            this.rootNodeRenderer.destroy(device);
-    }
-}
-
-const bindingLayouts: GfxBindingLayoutDescriptor[] = [
-    { numUniformBuffers: 3, numSamplers: 2, },
-];
 
 export class DK64Renderer implements Viewer.SceneGfx {
     public renderHelper: GfxRenderHelper;
@@ -1138,30 +255,6 @@ function decompress(buffer: ArrayBufferSlice): ArrayBufferSlice {
     assert(view.getUint32(0x00) === 0x1F8B0800);
     const decompressed = Deflate.decompress_raw(buffer.slice(0x0A));
     return decompressed;
-}
-
-interface SpriteData {
-    address: number;
-    id: number;
-    imagesPerFrameHorizontal: number;
-    imagesPerFrameVertical: number;
-    flags: number;
-    codec: number;
-    params: number[];
-    table: number;
-    width: number;
-    height: number;
-    images: number[];
-}
-
-interface EnvironmentParticleData {
-    map: number;
-    start: [number, number, number];
-    end: [number, number, number];
-    gap: number;
-    distance: number;
-    baseScale: number;
-    risingScale: number;
 }
 
 class TextureData {
@@ -1345,181 +438,6 @@ function addSceneActors(
             renderer.setPositionYAnimation(definition.positionYAmplitude * worldScale, definition.rotationYSpeed ?? 0);
         renderer.setCullBoundingBox(renderer.computeWorldBoundingBox());
         sceneRenderer.meshRenderers.push(renderer);
-    }
-}
-
-interface SpriteParticleEvent {
-    origin: vec3;
-    spawnTick: number;
-    frameOffset?: number;
-    velocityY?: number;
-}
-
-function addSpriteParticleEvents(device: GfxDevice, cache: GfxRenderCache, sceneRenderer: DK64Renderer, sharedOutput: RSPSharedOutput, romData: ROMData, definition: SpriteData, events: SpriteParticleEvent[], scale: number, loopTicks: number, frameDuration = 1, lifetime: number | undefined = definition.images.length * frameDuration, color: readonly number[] = definition.params.slice(0, 4), maxDistance?: number, fadeStartDistance?: number): void {
-    const sourceTable = definition.table !== 0 ? romData.TexData : romData.AnimTexData;
-    const sourceFrames = definition.images.map((image) => sourceTable[image]);
-    const frames = sourceFrames.flatMap((frame) => new Array(frameDuration).fill(frame));
-    const animationTickCount = frames.length;
-
-    for (let phase = 0; phase < animationTickCount; phase++) {
-        const phaseEvents = events.filter((event) => {
-            const requestedTick = (event.frameOffset ?? 0) * frameDuration;
-            const animationOffset = ((requestedTick - event.spawnTick) % animationTickCount + animationTickCount) % animationTickCount;
-            return animationOffset === phase;
-        });
-        for (let eventBase = 0; eventBase < phaseEvents.length; eventBase += 8) {
-            const batch = phaseEvents.slice(eventBase, eventBase + 8);
-            const segment = 0x0E;
-            const segmentBuffers: ArrayBufferSlice[] = [];
-            segmentBuffers[0x08] = createSpriteVertexBuffer(definition, batch.length);
-            const animation: AnimatedTexture[] = [{
-                segment,
-                // Ensure sprites on the same segment+phase don't overlap in the texture cache.
-                group: definition.id,
-                frameDuration: 1,
-                frameOffset: phase,
-                frames,
-            }];
-            const state = new RSPState(romData.TexData, segmentBuffers, sharedOutput, animation);
-            initDL(state, false);
-            initSpriteMaterial(state, definition, segment, color);
-            const firstVertex = sharedOutput.vertices.length;
-            for (let quad = 0; quad < batch.length; quad++) {
-                state.gSPVertex(0x08000000 + quad * 4 * 0x10, 4, 0);
-                state.gSPTri(0, 1, 2);
-                state.gSPTri(0, 2, 3);
-            }
-            const output = state.finish()!;
-
-            const width = definition.width * scale * 3;
-            const height = definition.height * scale * 3;
-            const mesh: Mesh = {
-                sharedOutput,
-                rspState: state,
-                rspOutput: output,
-                spriteBillboards: batch.map((event, index) => new SpriteBillboard(
-                    firstVertex + index * 4,
-                    event.origin,
-                    width / 2,
-                    height / 2,
-                    {
-                        spawnTick: lifetime === undefined ? undefined : event.spawnTick,
-                        lifetime,
-                        loopTicks: lifetime === undefined ? undefined : loopTicks,
-                        velocityY: event.velocityY,
-                        maxDistance,
-                        fadeStartDistance,
-                    },
-                )),
-            };
-            const meshData = new MeshData(device, cache, mesh);
-            sceneRenderer.meshDatas.push(meshData);
-            const renderer = new RootMeshRenderer(device, cache, meshData, SceneRenderLayer.Effects, sceneRenderer.fogParams, sceneRenderer.gpuTextureCache);
-            renderer.sortKeyBase = makeSortKey(GfxRendererLayer.TRANSLUCENT);
-            renderer.setBackfaceCullingEnabled(false);
-            sceneRenderer.meshRenderers.push(renderer);
-        }
-    }
-}
-
-function isAlwaysRunningInitialBlock(block: ScriptBlock): boolean {
-    return block.conditions.length === 1
-        && block.conditions[0].opcode === 1
-        && block.conditions[0].args[0] === 0
-        && !block.executions.some((command) => command.opcode === 1);
-}
-
-function nextEffectRandom(state: { value: number }): number {
-    // The game uses its own RNG, this isolated RNG is simpler.
-    state.value = (Math.imul(state.value, 0x41C64E6D) + 0x3039) >>> 0;
-    return state.value >>> 16;
-}
-
-function interpolateEnvironmentParticle(entry: EnvironmentParticleData, offset: number): vec3 {
-    return vec3.fromValues(
-        (entry.start[0] + (entry.end[0] - entry.start[0]) * offset) * 3,
-        (entry.start[1] + (entry.end[1] - entry.start[1]) * offset) * 3,
-        (entry.start[2] + (entry.end[2] - entry.start[2]) * offset) * 3,
-    );
-}
-
-function addEnvironmentalEffects(device: GfxDevice, cache: GfxRenderCache, sceneRenderer: DK64Renderer, sharedOutput: RSPSharedOutput, romData: ROMData, map: DK64Map, mapID: number, props: readonly SetupProp[], scripts: InstanceScript[]): void {
-    const propsByID = new Map(props.map((prop) => [prop.id, prop]));
-    const spriteByAddress = new Map(romData.SpriteData.map((sprite) => [sprite.address, sprite]));
-    const loopTicks = 900;
-
-    // from func_global_asm_80664CB0 + func_global_asm_80664D20:
-    // ambient waterfall emitters.
-    const baseSpray = spriteByAddress.get(0x8072140C);
-    const risingSpray = spriteByAddress.get(0x8071FF18);
-    if (baseSpray !== undefined && risingSpray !== undefined) {
-        for (const [entryIndex, entry] of romData.EnvironmentParticleData.entries()) {
-            if (entry.map !== mapID)
-                continue;
-            const random = { value: (mapID << 16) ^ entryIndex ^ 0x664D20 };
-            const baseEvents: SpriteParticleEvent[] = [];
-            const risingEvents: SpriteParticleEvent[] = [];
-            for (let tick = 0; tick < loopTicks; tick++) {
-                if (tick % 18 === 0) {
-                    for (let offset = 0; offset <= 1.00001; offset += entry.gap) {
-                        baseEvents.push({
-                            origin: interpolateEnvironmentParticle(entry, Math.min(offset, 1)),
-                            spawnTick: tick,
-                            frameOffset: (nextEffectRandom(random) % 10000) % 6,
-                        });
-                    }
-                }
-                if (tick % 10 === 0) {
-                    risingEvents.push({
-                        origin: interpolateEnvironmentParticle(entry, ((nextEffectRandom(random) % 10000) % 1000) / 1000),
-                        spawnTick: tick,
-                        velocityY: 1.7 * 3,
-                    });
-                }
-            }
-            const drawDistance = entry.distance * 3;
-            // from func_global_asm_80717B64: particle spawn alpha control
-            addSpriteParticleEvents(device, cache, sceneRenderer, sharedOutput, romData, baseSpray, baseEvents, entry.baseScale, loopTicks, 3, 18, [0xFF, 0xFF, 0xFF, 0x96], drawDistance, drawDistance * 3 / 4);
-            addSpriteParticleEvents(device, cache, sceneRenderer, sharedOutput, romData, risingSpray, risingEvents, entry.risingScale, loopTicks, 3, 30, [0xFF, 0xFF, 0xFF, 0x96], entry.distance * 3);
-        }
-    }
-
-    for (const script of scripts) {
-        const prop = propsByID.get(script.id);
-        if (prop === undefined)
-            continue;
-        for (const block of script.blocks) {
-            if (!isAlwaysRunningInitialBlock(block))
-                continue;
-            for (const command of block.executions) {
-                if (command.opcode !== 7)
-                    continue;
-                const functionAddress = romData.CustomScriptFunctionData[command.args[0]];
-
-                // from func_global_asm_80644EC8: 2 sprites (D_global_asm_80720A7C) per tick
-                if (functionAddress === 0x80644EC8) {
-                    const definition = spriteByAddress.get(0x80720A7C)!;
-                    const frequency = command.args[1];
-                    const requestedPointCount = command.args[2];
-                    const random = { value: (mapID << 16) ^ script.id ^ 0x44EC8 };
-                    const events: SpriteParticleEvent[] = [];
-                    for (let tick = 0; tick < loopTicks; tick++) {
-                        for (let set = 0; set < 2; set++) {
-                            if ((nextEffectRandom(random) % frequency) !== 0)
-                                continue;
-                            const points = map.effectPointSets[set]!;
-                            const pointCount = Math.min(requestedPointCount, points.length);
-                            const point = points[nextEffectRandom(random) % pointCount];
-                            events.push({
-                                origin: vec3.fromValues(point[0] * 3, point[1] * 3, point[2] * 3),
-                                spawnTick: tick,
-                            });
-                        }
-                    }
-                    addSpriteParticleEvents(device, cache, sceneRenderer, sharedOutput, romData, definition, events, 1.2, loopTicks);
-                }
-            }
-        }
     }
 }
